EXTANG-117 Issue of URLSearchParams is eradicated.
